+++ Admin-Service/src/main/java/com/adm/app/AdminServiceApplication.java
@@ -1,0 +1,15 @@
+package com.adm.app;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
+
+@SpringBootApplication
+@EnableEurekaClient
+public class AdminServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(AdminServiceApplication.class, args);
+	}
+
+}
