+++ Admin-Service/src/main/java/com/adm/app/model/EmployeeAddress.java
@@ -1,0 +1,26 @@
+package com.adm.app.model;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Entity
+@AllArgsConstructor
+@NoArgsConstructor
+public class EmployeeAddress {
+
+	@Id
+	@GeneratedValue(strategy = GenerationType.AUTO)
+	private int eAddId;
+	private String eState;
+	private String eCity;
+	private String eLandmark;
+	private int pincode;
+	
+}
