+++ Admin-Service/src/main/java/com/adm/app/repository/AdminRepository.java
@@ -1,0 +1,14 @@
+package com.adm.app.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;	
+import org.springframework.stereotype.Repository;
+
+import com.adm.app.model.Employee;
+
+@Repository
+public interface AdminRepository extends JpaRepository<Employee, Integer> {
+
+//	void saveEmployeeByempId(int id);
+
+	
+}
