+++ Admin-Service/src/main/resources/application.properties
@@ -1,0 +1,2 @@
+spring.config.import=optional:configserver:http://localhost:9090
+spring.profiles.active=Admin
