+++ Admin-Service/src/test/java/com/adm/AdminServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.adm;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class AdminServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
