--- Admin/.settings/org.springframework.ide.eclipse.prefs
@@ -1,2 +1,0 @@
-boot.validation.initialized=true
-eclipse.preferences.version=1
