--- Admin/src/main/java/com/admin/app/repository/AdminRepository.java
@@ -1,14 +1,0 @@
-package com.admin.app.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import com.admin.app.model.Employee;
-
-@Repository
-public interface AdminRepository extends JpaRepository<Employee, Integer> {
-
-//	void saveEmployeeByempId(int id);
-
-	
-}
