--- Admin/src/main/java/com/admin/app/service/AdminService.java
@@ -1,17 +1,0 @@
-package com.admin.app.service;
-
-
-import org.springframework.http.ResponseEntity;
-
-import com.admin.app.model.Employee;
-
-public interface AdminService {
-
-	Employee addemployee(Employee employee);
-
-	ResponseEntity<?> getEmployeeById(int empId);
-
-    void deleteEmployee(int id);
-
-
-}
