--- Admin/src/main/resources/application.properties
@@ -1,2 +1,0 @@
-spring.config.import=optional:configserver:http://localhost:9090
-spring.profiles.active=Admin
